Code with the three layer neural network and cross validation
This code shows that a 3-layer model has the better accuracy than the
submitted 2-layer on the training dataset. However, it has smaller
accuracy score on the test set, which suggests it is over-fitting and
thus, has the inferior generalization performance. This is confirmed by
comparing the 5-fold cross-validation accuracy scores—the 2-layer model
has better scores.